Fix TWCmdGenTemporalDataModule where it was using the wrong global graph
Before this commit, TWCmdGenTemporalDataModule was using the global
graph generated by TWCmdGenTemporalDataset. Unfortunately, this graph is
wrong as it is only the final graph after going through all the graph
events, but TWCmdGenTemporalDataModule requires intermediate graphs as
well. In order to support this, TWCmdGenTemporalDataModule now builds
its own global graph as it goes through graph events.