Revert "Use node masks in EventNodeHead"
This reverts commit 341e42d6d01c74b4112813154009a88d373b5ccf.

We shouldn't be zero masking logits as it can actually make the padded
logit the highest when other logits are negative. As a result, the
correct time to mask is after softmax. However, we should be able to do
away with masking as the model should learn to ignore the pad node.