Left-shift graph events returned by greedy decoding
We were returning the generated graph event that comes after the graph
data, which caused problems in evaluation, so left-shift all the graph
events returned by greedy decoding. This means that greedy decoding will
return a "start" event and drop the last event. This is OK since the
start event doesn't have any effect on graphs and the last event is
supposed to be an "end" event anyway.