Manually append new nodes when updating batched graphs
tensor.sort() is not stable on GPUs, so the order of the nodes changes
randomly after adding new nodes. This is problematic b/c we follow a
specific order (new nodes always go to the back in the list of nodes in
a graph, which is the default behavior of networkx) when generating
training data. So we do that operation manually using tensor operations.