Stop using batch.num_graphs
When there are empty graphs at the end of a batch, batch.num_graphs will
not count them unless it was created via Data.from_data_list(), which
manually sets batch._num_graphs. This means that batch.num_graphs is not
necessarily the same as the batch size in our case.