Adjust batch sizes and learning rate
Now that we use the word embeddings with a learned linear layer for
label embeddings, the GPU memory usage has gone up, so halve the
training batch size and learning rate. Double the validation and test
batch sizes since we now have a bit more room in memory during training.