Separate out validation and test metrics modules
If you use the same metrics module for the validation loop and test
loop, the test loop metrics will be skewed b/c the metrics collected in
the last validation loop that has been interrupted is not reset. The
correct way to handle these is to have separate metrics modules for the
validation loop and test loop.