Add hugging face text encoder parity tests
These parity tests ensure that the outputs for ID-based and
one-hot-encoded inputs to the hugging face text encoders are identical.

We also remove support for DistilBERT as it handles inputs_embeds
incorrectly, i.e. it assumes that inputs_embeds already has positional
embeddings added to it. We can handle this, but we'll do it if we
actually need to.